Script to work on frenzzup.com.
Some control ID/Class has been changed on new site frenzzup.com. Updated the script to work on that. copy it and use the same way as before.
One change which it will show is, will not show total links to click on site. Rather only show the count of clicks which are not clicked yet. So if out of 100 clicks, 20 has been clicked and you run the script, it will tell 80 clicks are there. So do not be bothered, it is going to click all the links which are not clicked yet.