Script to work on frenzzup.com.
Some control ID/Class has been changed on new site frenzzup.com. Updated the script to work on that. copy it and use the same way as before.
One change which it will show is, will not show total links to click on site. Rather only show the count of clicks which are not clicked yet. So if out of 100 clicks, 20 has been clicked and you run the script, it will tell 80 clicks are there. So do not be bothered, it is going to click all the links which are not clicked yet.
Also note that on some browser's, if you open a new tab, the script seems like its not working. That is not a problem with the script but the JavaScript timer which goes to low priority when a new tab is opened. This is done by the browser's and so it looks like script has stopped. Don't bother, just do one click manually and the script will again come to life and start doing the work! No need to start the script again in that scenario.