More updates to the script for frenzzup.com
Some links when clicked throw's exception and do not show as clicked. So the script keep on clicking those links again and again till they show as clicked. This issue was making the script blocked after every few clicks... Seems to be working fine now. Also the number of clicks on the page by script is updated. It shows the correct number of links on the page and currently clicked link's index as well.